#209 Editor to typescript, update typings
