+++ packages/ketcher-react/src/script/editor/index.ts
@@ -1,0 +1,4 @@
+import { Editor } from './Editor'
+
+export { Editor }
+export default Editor
