+++ packages/ketcher-react/src/script/editor/utils/elementOffset.ts
@@ -1,0 +1,30 @@
+/****************************************************************************
+ * Copyright 2021 EPAM Systems
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ ***************************************************************************/
+
+// TODO: find DOM shorthand
+import Vec2 from '../../util/vec2'
+
+export function elementOffset(element) {
+  let top = 0
+  let left = 0
+  do {
+    top += element.offsetTop || 0
+    left += element.offsetLeft || 0
+    element = element.offsetParent
+  } while (element)
+
+  return new Vec2(left, top)
+}
